Rename to protobuf, add internal.proto
