+++ encoding/x/protobuf/doc.go
@@ -18,7 +18,7 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
 
-// Package pb implements Protocol Buffers encoding support for YARPC.
+// Package protobuf implements Protocol Buffers encoding support for YARPC.
 //
 // https://developers.google.com/protocol-buffers/docs/proto3
-package pb
+package protobuf
